ZugINI.xml file gets replaced only if it don't exist.

+++ trunk/src/main/resources/InstallZugLinux.sh
@@ -1,5 +1,8 @@
 sudo mv /usr/lib/automature/ZUG/zug.linux /usr/lib/automature/ZUG/zug
+if [ ! -e /usr/lib/automature/ZUG/ZugINI.xml ]
+ then
 sudo mv /usr/lib/automature/ZUG/ZugINI.xml.Linux /usr/lib/automature/ZUG/ZugINI.xml
+fi
 sudo rm /usr/lib/automature/ZUG/ZugINI.xml.Mac
 sudo rm /usr/lib/automature/ZUG/ZugINI.xml.Windows
 sudo chmod 755 /usr/lib/automature/ZUG/zug
